Add IsReplicationFactorUpdating method to Client
Before we introduced the in-place replication.factor changes, the
replication.factor of a topic was updated atomically across all of its
partitions:

  * when the topic was first created, all of its initial partitions
    had the same rf

  * when partitions were added to a topic, they would be replicated
    according to the topic's rf

  * when a topic's rf was changed, it would be deleted and created
    again (bullet point 1)

Now that a topic's rf may be updated in-place, the rf is not updated
atomically; it's instead updated in a rolling/eventually-consistent
manner. For sometime after the apply, some partitions will have the
new rf, and some will still have the old rf. Eventually, however,
they'll all have the new rf.

This commit introduces a new method to determine whether or not an rf
update has finished propagating through all of a topic's partitions,
and changes the kafka_topic resource to be aware of this
propagating/updating state.