Wait for rf update before adding partitions
We can't add partitions while a replication.factor update is ongoing,
so this commit waits for an rf update to finish before adding any new
partitions.

Because we're now waiting for the rf to update separately, the
IsReplicationFactorUpdating logic has been removed from
topicRefreshFunc (which only waits for partition and config updates to
finish).