Add CanAlterReplicationFactor method
This new method determines if the cluster supports altering a topic's
replication.factor in-place. If so, then the provider will go ahead
and make alterations in-place; otherwise, it will replace the topic
like it did before.

populateAPIVersions was changed to ensure that the max api versions
are supported by every broker in the cluster.