Add support for importing kafka_acl resources
Importer is set in a way, that if we can reconstruct the data just from
Id(), we may just use it.

This PR accomplishes that.
It also adds documentation on how to do import and reorders fields to
match the import format - so that one won't be confused so easily.